python/requirements.txt
# core
attrs
sympy

# exteral AD packages
autograd
jax
jaxlib
torch

# testing and development
black
flake8
isort
mypy
pylint
pytest-cov
